osc.c: Update to latest iio_buffer API from libiio
When using iio_buffer objects, it is no more needed to open/close
the device, as it is done internally. Plus, the iio_device_create_buffer
now takes the buffer size in samples count and not in bytes.

It also takes a new parameter, a boolean to instruct whether or not
the iio_buffer is an output buffer.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>